drm/panel: Add support for In-Cell Synaptics/JDI H455TAX01 combo
This adds support for the In-Cell Synaptics TS controller plus
JDI Display H455TAX01.0 combo assembly.
It is unclear and nowhere to be found whether this model is
relative to the display itself or to the entire assembly.

This is a 720x1280@60 display on MIPI DSI, found in some Sony
Xperia X Compact smartphones.

Signed-off-by: Angelo G. Del Regno <kholk11@gmail.com>