clk: qcom: gcc-8976: Sort-of-unnecessary hacks for the current state
The GPU GX needs to be hacked in as a parent of CX because there is
no support for multiple power domains from DT and we don't have any
regulator nor power-domain controller that will bring up the CX one
separately from the GX one for us.... FOR NOW :))

The idea is to either get the GPU driver to use multiple genpds or,
better, to bring up the CX/GX GDSC(s) in the CPR regulator driver
that will "provide power to the GPU". This will be done later.