kvfree_rcu(): Switch from kmalloc/kfree to __get_free_page/free_page.
The advantages of using kmalloc() and kfree() are a possible small speedup
on CONFIG_SLAB=y systems, avoiding the allocation-side cast, and use of
more-familiar API members.  The advantages of using __get_free_page()
and free_page() are a possible reduction in fragmentation and direct
access to the buddy allocator.

To help settle the question as to which to use, this commit switches
from kmalloc() and kfree() to __get_free_page() and free_page().

Suggested-by: Michal Hocko <mhocko@suse.com>
Suggested-by: "Uladzislau Rezki (Sony)" <urezki@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>